Adding assert for clean project dirs
