+++ workspaces/templates-management/packages/project-build/src/projectBuild.ts
@@ -1,16 +1,14 @@
 import { S3TemplateRepository } from '@goldstack/template-repository';
 import extract from 'extract-zip';
-import { rm, mkdir, write, read } from '@goldstack/utils-sh';
+import { rm, write, read } from '@goldstack/utils-sh';
 import { AssertionError } from 'assert';
 import { debug } from '@goldstack/utils-log';
 import path, { resolve } from 'path';
-import {
-  ProjectConfiguration,
-  PackageProjectConfiguration,
-} from '@goldstack/utils-project';
+import { ProjectConfiguration } from '@goldstack/utils-project';
 
 import { readPackageConfig } from '@goldstack/utils-package';
 import { readdirSync } from 'fs';
+import { buildTemplate } from './buildTemplate';
 
 export interface TemplateReference {
   name: string;
@@ -23,7 +21,7 @@ export interface ProjectBuildParams {
   destinationDirectory: string;
 }
 
-const assertTemplateReferenceVersion = async (
+export const assertTemplateReferenceVersion = async (
   s3: S3TemplateRepository,
   templateReference: TemplateReference
 ): Promise<TemplateReference> => {
@@ -46,7 +44,7 @@ const assertTemplateReferenceVersion = async (
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-function assert(condition: any, msg?: string): asserts condition {
+export function assert(condition: any, msg?: string): asserts condition {
   if (!condition) {
     throw new AssertionError({ message: msg });
   }
@@ -66,7 +64,10 @@ function sortKeys(obj: any): any {
   return obj; // Return the value if it's not an object or array
 }
 
-const setPackageName = (packageFolder: string, packageName: string): void => {
+export const setPackageName = (
+  packageFolder: string,
+  packageName: string
+): void => {
   const goldstackPackageConfig = readPackageConfig(packageFolder);
   goldstackPackageConfig.name = packageName;
   write(
@@ -84,54 +85,11 @@ const setPackageName = (packageFolder: string, packageName: string): void => {
   );
 };
 
-const buildTemplate = async (
-  params: ProjectBuildParams,
-  packageConfig: PackageProjectConfiguration
-): Promise<void> => {
-  debug(
-    `Building package ${packageConfig.packageName} in ${params.destinationDirectory}`
-  );
-  const template: TemplateReference = {
-    name: packageConfig.templateReference.templateName,
-    version: packageConfig.templateReference.templateVersion,
-  };
-
-  const templateReference = await assertTemplateReferenceVersion(
-    params.s3,
-    template
-  );
-
-  assert(templateReference.version);
-
-  const packageFolder = path.join(
-    params.destinationDirectory,
-    'packages',
-    `${packageConfig.packageName}/`
-  );
-
-  mkdir('-p', packageFolder);
-  const zipPath = await params.s3.downloadTemplateArchive(
-    templateReference.name,
-    templateReference.version,
-    packageFolder
-  );
-
-  assert(zipPath);
-  await extract(zipPath, { dir: path.resolve(packageFolder) });
-
-  rm('-f', zipPath);
-  debug('Template archive extracted to ' + path.resolve(packageFolder), {
-    packageFolder: path.resolve(packageFolder),
-    filesInPackageFolder: readdirSync(path.resolve(packageFolder)).join(', '),
-  });
-  setPackageName(packageFolder, packageConfig.packageName);
-};
-
 export const buildProject = async (
   params: ProjectBuildParams
 ): Promise<void> => {
   debug(
-    `Building project ${params.config.projectName} from to ${params.destinationDirectory}`
+    `Building project ${params.config.projectName} into ${params.destinationDirectory}`
   );
   const config = params.config;
